rename: EmbedMessage -> embed-message
